Rename knex-migrator
Rename lib/model/migrate.js to lib/model/knex-migrator.js.

This is preparation for deprecating the knex migrator and replacing it in #1363.
