+++ lib/bin/check-open-db-queries.js
@@ -7,7 +7,7 @@
 // including this file, may be copied, modified, propagated, or distributed
 // except according to the terms contained in the LICENSE file.
 
-const { withKnex } = require('../model/migrate');
+const { withKnex } = require('../model/knex-migrator');
 
 (async () => {
   try {
