+++ lib/bin/run-migrations.js
@@ -7,7 +7,7 @@
 // including this file, may be copied, modified, propagated, or distributed
 // except according to the terms contained in the LICENSE file.
 
-const { withKnex, migrate } = require('../model/migrate');
+const { withKnex, migrate } = require('../model/knex-migrator');
 
 (async () => {
   try {
